requirements.txt
# streamlit==1.29.0
# pandas==2.1.4
# groq==0.11.0
# langchain>=0.1.0
# langchain-groq==0.1.9
# python-dotenv==1.0.0
# pydantic==2.5.3
# pyyaml==6.0.1
# plotly==5.18.0
# langgraph==0.0.15
# typing_extensions==4.9.0
# httpx==0.27.2

streamlit==1.29.0
pandas==2.1.4
groq>=0.37.1
langchain>=1.2.0
langchain-groq>=1.1.1
langchain-community>=0.4.1
python-dotenv==1.0.0
pydantic>=2.5.3
pyyaml==6.0.1
plotly==5.18.0
langgraph>=1.0.5
typing_extensions>=4.9.0
packaging<24,>=16.8